Moved type declarations to NamespaceResolver
