+++ packages/cli/src/languagePlugins/csharp/namespaceResolver/index.ts
@@ -1,6 +1,24 @@
 import Parser from "tree-sitter";
-import { File, Namespace, ExportedSymbol } from "../types";
 import { csharpParser } from "../../../helpers/treeSitter/parsers";
+
+export interface File {
+  path: string;
+  rootNode: Parser.SyntaxNode;
+}
+
+export interface ExportedSymbol {
+  name: string;
+  type: string;
+  node: Parser.SyntaxNode;
+  namespace?: string;
+  filepath: string;
+}
+
+export interface Namespace {
+  name: string;
+  exports: ExportedSymbol[];
+  childrenNamespaces: Namespace[];
+}
 
 export class NamespaceResolver {
   parser: Parser = csharpParser;
